requirements.txt
pandas
numpy
pickle
streamlit
python-dotenv
requests
scikit-learn